separated progess into files
