+++ src/progress/checklist.js
@@ -1,59 +1,5 @@
 import { read_json, write_json, delete_json, mkdir } from './utils/files.js'
 
-/* this class make a list that is saved disk, and or read from */
-class DiskList{
-    constructor(name, values = null, path = null){
-        this.dir_path = path? path : '../data/resources/list/';
-        this.name = name + ".json";
-        // try to read already saved values 
-        if(values){ // if values have be passed
-            this.values = values;
-            write_json(this.values, this.dir_path + this.name);
-        }else // try to read from disk
-            this.values = read_json( this.dir_path + this.name) ?? [];
-    }
-    // save value
-    add = value => {
-        this.values.push(value);
-        return write_json(this.values, this.dir_path + this.name);
-    }
-}
-
-/* this class is similar to the disklist, but it remove any repeated values */
-class DiskSet{
-    constructor(name, values = null, path){
-        this.dir_path = path?? '../data/resources/list/';
-        this.name = name + ".json";
-        this.set = new Set();
-        this.array = [];
-        // try to read already saved values
-        if(!values)
-            values = read_json( this.dir_path + this.name) ?? [];
-        // check uniquenes
-        for (var value of values) this._add(value);
-        // after done checking save to memeory
-        this._save();
-    }
-
-    // save value
-    _save = () => write_json( this.array, this.dir_path + this.name );
-
-    // add value to set, in unique it add it to array
-    _add = value => {
-        if( this.set.has(JSON.stringify(value)) )
-            return false;
-        else{ // if it not in the set
-            this.set.add(JSON.stringify(value));
-            this.array.push(value);
-            return true;
-        }
-    }
-
-    // add and saves value
-    add = value => (this._add(value))? this._save() : false ;
-}
-
-
 /* this class makes a checklist for value that need to be check,
  * it takes a check function whihc goes throught the values. */
 class Checklist{
@@ -61,6 +7,7 @@ class Checklist{
     constructor(name, values, path){
         // only for script
         this.dir_path =  path ?? './data/resources/checklists/';
+        mkdir(this.dir_path);
         this.name = name + ".json";
         this.filename = this.dir_path + this.name
         this.checklist = read_json( this.filename );
@@ -170,8 +117,6 @@ class Checklist{
         this.checklist = []
         delete_json(this.dir_path + this.name)
     }
-
-
 }
 
-export { Checklist, DiskList, DiskSet }
+export default Checklist;
